Revamp SyncConsole
- Be more specific in Console's scaladoc, add examples
- Move SyncConsole to a new file
- Make input/output streams of SyncConsole customizable
- Make SyncConsole private, provide factories
