+++ core/src/main/scala/cats/effect/Console.scala
@@ -16,70 +16,14 @@
 
 package cats.effect
 
-import cats.{ ~>, Show }
 import cats.instances.string._
-import cats.syntax.show._
+import cats.{ ~>, Show }
 
 /**
   * Effect type agnostic `Console` with common methods to write and read from the standard console.
-  */
-trait Console[F[_]] {
-
-  /**
-    * Prints a message of type A to the standard console followed by a new line if an instance `Show[A]` is found.
-    */
-  def putStrLn[A: Show](a: A): F[Unit]
-
-  /**
-    * Prints a message to the standard console followed by a new line.
-    */
-  def putStrLn(str: String): F[Unit] = putStrLn[String](str)
-
-  /**
-    * Prints a message of type A to the standard console if an instance `Show[A]` is found.
-    */
-  def putStr[A: Show](a: A): F[Unit]
-
-  /**
-    * Prints a message to the standard console.
-    */
-  def putStr(str: String): F[Unit] = putStr[String](str)
-
-  /**
-    * Prints a message of type A to the standard error output stream if an instance `Show[A]` is found.
-    */
-  def putError[A: Show](a: A): F[Unit]
-
-  /**
-    * Prints a message to the standard error output stream.
-    */
-  def putError(str: String): F[Unit] = putError[String](str)
-
-  /**
-    * Reads line from the standard console.
-    *
-    * @return a value representing the user's input or raise an error in the F context.
-    */
-  def readLn: F[String]
-
-  /**
-    * Transforms this console using a FunctionK.
-    * */
-  def mapK[G[_]](f: F ~> G): Console[G] = new TransformedConsole(this, f)
-}
-
-object Console {
-
-  def apply[F[_]](implicit ev: Console[F]): Console[F] = ev
-
-  /**
-    * Default instance for `Console[IO]`
-    */
-  object io extends SyncConsole[IO]
-}
-
-/**
-  * A default instance for any `F[_]` with a `Sync` instance.
+  *
+  * For an implementation that uses standard input/output,
+  * see [[Console.io]] (for `Console[IO]`) or [[SyncConsole.stdio]] (for any `F[_]: Sync`).
   *
   * Use it either in a DSL style:
   *
@@ -96,7 +40,7 @@ object Console {
   *     } yield ()
   * }}}
   *
-  * Or in tagless final encoding:
+  * Or in Tagless Final encoding:
   *
   * {{{
   *   import cats.Monad
@@ -110,19 +54,58 @@ object Console {
   *            else C.putError("Name is empty!")
   *     } yield ()
   * }}}
-  *
   */
-class SyncConsole[F[_]](implicit F: Sync[F]) extends Console[F] {
-  def putStrLn[A: Show](a: A): F[Unit] =
-    F.delay(println(a.show))
-  def putStr[A: Show](a: A): F[Unit] =
-    F.delay(print(a.show))
-  def putError[A: Show](a: A): F[Unit] =
-    F.delay(System.err.println(a.show))
-  def readLn: F[String] =
-    F.delay(scala.io.StdIn.readLine)
-  // Needed to avoid "ambiguous reference to overloaded definition" error
-  override final def putStrLn(str: String): F[Unit] = super.putStrLn(str)
-  override final def putStr(str: String): F[Unit]   = super.putStr(str)
-  override final def putError(str: String): F[Unit] = super.putError(str)
+trait Console[F[_]] {
+
+  /**
+    * Prints a message of type A to the console followed by a new line using the implicit `Show[A]` instance.
+    */
+  def putStrLn[A: Show](a: A): F[Unit]
+
+  /**
+    * Prints a message to the console followed by a new line.
+    */
+  def putStrLn(str: String): F[Unit] = putStrLn[String](str)
+
+  /**
+    * Prints a message of type A to the console using the implicit `Show[A]` instance.
+    */
+  def putStr[A: Show](a: A): F[Unit]
+
+  /**
+    * Prints a message to the console.
+    */
+  def putStr(str: String): F[Unit] = putStr[String](str)
+
+  /**
+    * Prints a message of type A to the error output using the implicit `Show[A]` instance.
+    */
+  def putError[A: Show](a: A): F[Unit]
+
+  /**
+    * Prints a message to the error output.
+    */
+  def putError(str: String): F[Unit] = putError[String](str)
+
+  /**
+    * Reads a line from the console input.
+    *
+    * @return a value representing the user's input.
+    */
+  def readLn: F[String]
+
+  /**
+    * Transforms this console using a FunctionK.
+    * */
+  def mapK[G[_]](f: F ~> G): Console[G] = new TransformedConsole(this, f)
+}
+
+object Console {
+
+  def apply[F[_]](implicit F: Console[F]): Console[F] = F
+
+  /**
+    * Default instance for `Console[IO]` that prints to standard input/output streams.
+    */
+  val io: Console[IO] = SyncConsole.stdio[IO]
 }
